Desktop/Agent_IA/requirements.txt
streamlit
anthropic
python-dotenv
fpdf2